fix(producer): reorganize code I
