+++ aether-producer/producer/db.py
@@ -16,209 +16,34 @@
 # specific language governing permissions and limitations
 # under the License.
 
+from datetime import datetime
+import signal
+import sys
+
+import gevent
 from gevent import monkey, sleep
+from gevent.event import AsyncResult
+from gevent.queue import PriorityQueue, Queue
+
 # need to patch sockets to make requests async
 monkey.patch_all()  # noqa
 import psycogreen.gevent
 psycogreen.gevent.patch_psycopg()  # noqa
 
-from datetime import datetime
-import logging
-import sys
-import os
-import signal
-import uuid
-
-import gevent
-from gevent.event import AsyncResult
-from gevent.queue import PriorityQueue, Queue
-
 import psycopg2
 from psycopg2 import sql
 from psycopg2.extras import DictCursor
 
-from sqlalchemy import Column, String
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import create_engine
+from sqlalchemy import Column, String, create_engine
 from sqlalchemy.exc import SQLAlchemyError
+from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import NullPool
 
-from producer.settings import Settings
-
-FILE_PATH = os.path.dirname(os.path.realpath(__file__))
+from producer.settings import SETTINGS, get_logger
 
 Base = declarative_base()
-logger = logging.getLogger('producer-db')
-logger.propagate = False
-handler = logging.StreamHandler()
-handler.setFormatter(logging.Formatter(
-    '%(asctime)s [ProducerDB] %(levelname)-8s %(message)s'))
-logger.addHandler(handler)
-engine = None
-Session = None
-
-file_path = os.environ.get('PRODUCER_SETTINGS_FILE')
-SETTINGS = Settings(file_path)
-log_level = logging.getLevelName(SETTINGS.get('log_level', 'DEBUG'))
-logger.setLevel(log_level)
-
-
-def init():
-    global engine
-
-    # Offset
-
-    offset_db_host = SETTINGS['offset_db_host']
-    offset_db_user = SETTINGS['offset_db_user']
-    offset_db_port = SETTINGS['offset_db_port']
-    offset_db_password = SETTINGS['offset_db_password']
-    offset_db_name = SETTINGS['offset_db_name']
-
-    url = f'postgresql+psycopg2://{offset_db_user}:{offset_db_password}' + \
-        f'@{offset_db_host}:{offset_db_port}/{offset_db_name}'
-
-    engine = create_engine(url, poolclass=NullPool)
-    try:
-        start_session(engine)
-        Offset.create_pool()
-        return
-    except SQLAlchemyError:
-        pass
-    try:
-        logger.error('Attempting to Create Database.')
-        create_db(engine, url)
-        start_session(engine)
-        Offset.create_pool()
-    except SQLAlchemyError as sqe:
-        logger.error('Database creation failed: %s' % sqe)
-        sys.exit(1)
-
-
-def start_session(engine):
-    try:
-        Base.metadata.create_all(engine)
-
-        global Session
-        Session = sessionmaker(bind=engine)
-
-        logger.info('Database initialized.')
-    except SQLAlchemyError as err:
-        logger.error('Database could not be initialized. | %s' % err)
-        raise err
-
-
-def create_db(engine, url):
-    db_name = url.split('/')[-1]
-    root = url.replace(db_name, 'postgres')
-    temp_engine = create_engine(root)
-    conn = temp_engine.connect()
-    conn.execute('commit')
-    conn.execute('create database %s' % db_name)
-    conn.close()
-
-
-def get_session():
-    return Session()
-
-
-def get_engine():
-    return engine
-
-
-def make_uuid():
-    return str(uuid.uuid4())
-
-
-class Offset(Base):
-
-    # SQLAlchemy is now only used for table creation.
-    # All gets and sets to offset are made via psycopg2 which
-    # allows us to use the priority connection pooling, making
-    # sure our sets happen ASAP while reads of offsets can be
-    # delayed until resources are free.
-
-    __tablename__ = 'offset'
-
-    # GET single offset value
-    GET_STR = '''
-        SELECT
-            o.schema_name,
-            o.offset_value
-        FROM public.offset o
-        WHERE o.schema_name = {schema_name}
-        LIMIT 1; '''
-
-    # UPSERT Postgres call.
-    SET_STR = '''
-        INSERT INTO public.offset
-            (schema_name, offset_value)
-        VALUES
-            ({schema_name}, {offset_value})
-        ON CONFLICT (schema_name) DO UPDATE SET
-            (schema_name, offset_value)
-            = (EXCLUDED.schema_name, EXCLUDED.offset_value);'''
-
-    schema_name = Column(String, primary_key=True)
-    offset_value = Column(String, nullable=False)
-
-    @classmethod
-    def create_pool(cls):
-        # Creates connection pool to OffsetDB as part of initialization.
-        offset_settings = {
-            'user': 'offset_db_user',
-            'dbname': 'offset_db_name',
-            'port': 'offset_db_port',
-            'host': 'offset_db_host',
-            'password': 'offset_db_password'
-        }
-        offset_creds = {k: SETTINGS.get(v) for k, v in offset_settings.items()}
-        global OFFSET_DB
-        offset_db_pool_size = SETTINGS.get('offset_db_pool_size', 6)
-        OFFSET_DB = PriorityDatabasePool(offset_creds, 'OffsetDB', offset_db_pool_size)
-
-    @classmethod
-    def update(cls, name, offset_value):
-        # Update or Create if not existing
-        call = 'Offset-SET'
-        try:
-            promise = OFFSET_DB.request_connection(0, call)  # Highest Priority
-            conn = promise.get()
-            cursor = conn.cursor()
-            query = sql.SQL(Offset.SET_STR).format(
-                schema_name=sql.Literal(name),
-                offset_value=sql.Literal(offset_value))
-            cursor.execute(query)
-            conn.commit()
-            return offset_value
-        except Exception as err:
-            raise err
-        finally:
-            try:
-                OFFSET_DB.release(call, conn)
-            except UnboundLocalError:
-                logger.error(f'{call} could not release a connection it never received.')
-
-    @classmethod
-    def get_offset(cls, name):
-        call = 'Offset-GET'
-        try:
-            promise = OFFSET_DB.request_connection(1, call)  # Lower Priority than set
-            conn = promise.get()
-            cursor = conn.cursor(cursor_factory=DictCursor)
-            query = sql.SQL(Offset.GET_STR).format(schema_name=sql.Literal(name))
-            cursor.execute(query)
-            res = [i for i in cursor]
-            if not res:
-                return None
-            return res[0][1]
-        except Exception as err:
-            raise err
-        finally:
-            try:
-                OFFSET_DB.release(call, conn)
-            except UnboundLocalError:
-                logger.error(f'{call} could not release a connection it never received.')
+logger = get_logger('producer-db')
 
 
 class PriorityDatabasePool(object):
@@ -234,8 +59,9 @@ class PriorityDatabasePool(object):
         self.name = name
         self.live_workers = 0
         self.pg_creds = pg_creds
-        self.max_connections = max_connections
+        self.max_connections = int(max_connections)
         logger.debug(f'Initializing DB Pool: {self.name} with {max_connections} connections.')
+
         self.workers = []
         self.backlog = 0
         self.max_backlog = 0
@@ -328,8 +154,152 @@ class PriorityDatabasePool(object):
                 c += 1
 
 
-# KernelDB
-pg_requires = ['user', 'dbname', 'port', 'host', 'password']
-pg_creds = {key: SETTINGS.get('postgres_%s' % key) for key in pg_requires}
-kernel_db_pool_size = SETTINGS.get('kernel_db_pool_size', 6)
-KERNEL_DB = PriorityDatabasePool(pg_creds, 'KernelDB', kernel_db_pool_size)  # imported from here
+class Offset(Base):
+
+    # SQLAlchemy is now only used for table creation.
+    # All gets and sets to offset are made via psycopg2 which
+    # allows us to use the priority connection pooling, making
+    # sure our sets happen ASAP while reads of offsets can be
+    # delayed until resources are free.
+
+    __tablename__ = 'offset'
+    schema_name = Column(String, primary_key=True)
+    offset_value = Column(String, nullable=False)
+
+    # private connection pool
+    __pool__ = None
+
+    # GET single offset value
+    GET_STR = '''
+        SELECT offset_value
+          FROM public.offset
+         WHERE schema_name = {schema_name}
+         LIMIT 1;
+    '''
+
+    # UPSERT Postgres call
+    SET_STR = '''
+        INSERT INTO public.offset
+            (schema_name, offset_value)
+        VALUES
+            ({schema_name}, {offset_value})
+        ON CONFLICT (schema_name)
+        DO UPDATE SET
+            (schema_name, offset_value) = (EXCLUDED.schema_name, EXCLUDED.offset_value);
+    '''
+
+    @classmethod
+    def create_pool(cls):
+        if not Offset.__pool__:
+            # Creates connection pool to OffsetDB as part of initialization.
+            offset_settings = {
+                'user': 'offset_db_user',
+                'dbname': 'offset_db_name',
+                'port': 'offset_db_port',
+                'host': 'offset_db_host',
+                'password': 'offset_db_password'
+            }
+            offset_creds = {k: SETTINGS.get(v) for k, v in offset_settings.items()}
+            offset_db_pool_size = SETTINGS.get('offset_db_pool_size', 6)
+            Offset.__pool__ = PriorityDatabasePool(offset_creds, 'OffsetDB', offset_db_pool_size)
+        return Offset.__pool__
+
+    @classmethod
+    def close_pool(cls):
+        if Offset.__pool__:
+            Offset.__pool__._kill()
+            Offset.__pool__ = None
+
+    @classmethod
+    def update(cls, name, offset_value):
+        # Update or Create if not existing
+        call = 'Offset-SET'
+        try:
+            promise = Offset.__pool__.request_connection(0, call)  # Highest Priority
+            conn = promise.get()
+            cursor = conn.cursor()
+            query = sql.SQL(Offset.SET_STR).format(
+                schema_name=sql.Literal(name),
+                offset_value=sql.Literal(offset_value),
+            )
+            cursor.execute(query)
+            conn.commit()
+            return offset_value
+        except Exception as err:
+            raise err
+        finally:
+            try:
+                Offset.__pool__.release(call, conn)
+            except UnboundLocalError:
+                logger.error(f'{call} could not release a connection it never received.')
+
+    @classmethod
+    def get_offset(cls, name):
+        call = 'Offset-GET'
+        try:
+            promise = Offset.__pool__.request_connection(1, call)  # Lower Priority than set
+            conn = promise.get()
+            cursor = conn.cursor(cursor_factory=DictCursor)
+            query = sql.SQL(Offset.GET_STR).format(schema_name=sql.Literal(name))
+            cursor.execute(query)
+            res = [i for i in cursor]
+            return res[0][0] if res else None
+        except Exception as err:
+            raise err
+        finally:
+            try:
+                Offset.__pool__.release(call, conn)
+            except UnboundLocalError:
+                logger.error(f'{call} could not release a connection it never received.')
+
+
+def init():
+    def _db_url(db_name):
+        return (
+            f'postgresql+psycopg2://{offset_db_user}:{offset_db_password}'
+            f'@{offset_db_host}:{offset_db_port}/{db_name}'
+        )
+
+    def _start_session(engine):
+        try:
+            Base.metadata.create_all(engine)
+            sessionmaker(bind=engine)
+            logger.info('Database initialized.')
+        except SQLAlchemyError as err:
+            logger.error('Database could not be initialized. | %s' % err)
+            raise err
+
+    def _create_db():
+        logger.info(f'Attempting to create database "{offset_db_name}".')
+        temp_engine = create_engine(_db_url('postgres'))
+        conn = temp_engine.connect()
+        conn.execute('commit')
+        conn.execute(f'create database {offset_db_name}')
+        conn.close()
+        logger.info(f'Database "{offset_db_name}" created.')
+
+    # Offset
+    offset_db_host = SETTINGS['offset_db_host']
+    offset_db_user = SETTINGS['offset_db_user']
+    offset_db_port = SETTINGS['offset_db_port']
+    offset_db_password = SETTINGS['offset_db_password']
+    offset_db_name = SETTINGS['offset_db_name']
+
+    engine = create_engine(_db_url(offset_db_name), poolclass=NullPool)
+    try:
+        _start_session(engine)
+        Offset.create_pool()
+        return
+    except SQLAlchemyError as sqe:
+        logger.error('Start session failed (1st attempt): %s' % sqe)
+        pass
+
+    # it was not possible to start session because the database does not exit
+    # create it and try again
+    try:
+        _create_db()
+        _start_session(engine)
+        Offset.create_pool()
+    except SQLAlchemyError as sqe:
+        logger.error('Start session failed (2nd attempt): %s' % sqe)
+        sys.exit(1)
