+++ aether-producer/producer/settings.py
@@ -16,8 +16,16 @@
 # specific language governing permissions and limitations
 # under the License.
 
+import enum
 import json
+import logging
 import os
+
+
+class KafkaStatus(enum.Enum):
+    SUBMISSION_PENDING = 1
+    SUBMISSION_FAILURE = 2
+    SUBMISSION_SUCCESS = 3
 
 
 class Settings(dict):
@@ -43,3 +51,51 @@ class Settings(dict):
             obj = json.load(f)
             for k in obj:
                 self[k] = obj.get(k)
+
+
+def get_logger(name, logger=None):
+    if not logger:
+        logger = logging.getLogger(name)
+        logger.propagate = False
+
+    try:
+        handler = logger.handlers[0]
+    except IndexError:
+        handler = logging.StreamHandler()
+        logger.addHandler(handler)
+
+    handler.setFormatter(logging.Formatter(f'%(asctime)s [{name}] %(levelname)-8s %(message)s'))
+
+    logger.setLevel(LOG_LEVEL)
+
+    return logger
+
+
+def _get_kafka_settings():
+    kafka_settings = {}
+    kafka_settings['bootstrap.servers'] = SETTINGS.get('kafka_url')
+
+    mode = SETTINGS.get('kafka_security', '')
+    if mode.lower() == 'sasl_plaintext':
+        # Let Producer use Kafka SU to produce
+        kafka_settings['security.protocol'] = 'SASL_PLAINTEXT'
+        kafka_settings['sasl.mechanisms'] = 'SCRAM-SHA-256'
+        kafka_settings['sasl.username'] = SETTINGS.get('KAFKA_SU_USER')
+        kafka_settings['sasl.password'] = SETTINGS.get('KAFKA_SU_PW')
+
+    elif mode.lower() == 'sasl_ssl':
+        kafka_settings['security.protocol'] = 'SASL_SSL'
+        kafka_settings['sasl.mechanisms'] = 'PLAIN'
+        kafka_settings['sasl.username'] = SETTINGS.get('KAFKA_SU_USER')
+        kafka_settings['sasl.password'] = SETTINGS.get('KAFKA_SU_PW')
+
+    return kafka_settings
+
+
+##################################################
+# Get settings from config file
+
+file_path = os.environ.get('PRODUCER_SETTINGS_FILE')
+SETTINGS = Settings(file_path)
+KAFKA_SETTINGS = _get_kafka_settings()
+LOG_LEVEL = logging.getLevelName(SETTINGS.get('log_level', 'DEBUG'))
